Scaling algorithm to end all scaling algorithms
Without any explicit 3d computations, compute the perspective
transformation mapping the quad to a 1x1 square using cross-ratios.
Map the given point to the square, then compute the norms of the images
of the unit x and y tangent vectors at that point, under the inverse
transformation. Do the same with the transformation defined by the frx,
fry, and frz and use the ratio of both as the scaling factor.