Add option to deproject shapes before applying
This means that the shape at the relative frame will not be changed
after applying, and all other shapes will follow the perspective track
relative to this relative shape.
This is done by computing the transformation that the transform tags on
the relative line apply, and inverting it for all shapes.
This will only work with \an7 shapes. To minimize rounding errors,
the origin of the shape should be near the tracked quad.